PWGHF: remove old DMesonPairs code and rename Testing code
