+++ PWGHF/HFC/DataModel/DMesonPairsTables.h
@@ -11,6 +11,7 @@
 
 /// \file DMesonPairsTables.h
 /// \brief D meson pair table definition.
+///
 /// \author Andrea Tavira García <tavira-garcia@ijclab.in2p3.fr>, IJCLab Orsay
 
 #ifndef PWGHF_HFC_DATAMODEL_DMESONPAIRSTABLES_H_
@@ -24,19 +25,18 @@ namespace o2::aod
 namespace hf_correlation_d_meson_pair
 {
 // Kinematic info
-DECLARE_SOA_COLUMN(DeltaPhi, deltaPhi, float); //! Delta phi of the pair
-DECLARE_SOA_COLUMN(DeltaEta, deltaEta, float); //! Delta eta of the pair
-DECLARE_SOA_COLUMN(PtCand1, ptCand1, float);   //! Transverse momentum of first candidate
-DECLARE_SOA_COLUMN(PtCand2, ptCand2, float);   //! Transverse momentum of second candidate
-DECLARE_SOA_COLUMN(YCand1, yCand1, float);     //! Rapidity of first candidate
-DECLARE_SOA_COLUMN(YCand2, yCand2, float);     //! Rapidity of second candidate
+DECLARE_SOA_COLUMN(PtCand1, ptCand1, float); //! Transverse momentum of first candidate
+DECLARE_SOA_COLUMN(PtCand2, ptCand2, float); //! Transverse momentum of second candidate
+DECLARE_SOA_COLUMN(YCand1, yCand1, float);   //! Rapidity of first candidate
+DECLARE_SOA_COLUMN(YCand2, yCand2, float);   //! Rapidity of second candidate
 // Invariant mass
-DECLARE_SOA_COLUMN(MCand1, mCand1, float); //! Invariant mass of first candidate
-DECLARE_SOA_COLUMN(MCand2, mCand2, float); //! Invariant mass of second candidate
-// Type of candidate: candidate charge and whether it is signal, reflected, or bkg
-DECLARE_SOA_COLUMN(CandidateType1, candidateType1, uint8_t); //! Type of first candidate
-DECLARE_SOA_COLUMN(CandidateType2, candidateType2, uint8_t); //! Type of second candidate
-DECLARE_SOA_COLUMN(DataType, dataType, uint8_t);             //! 0: data, 1: MC reco, 2: MC gen
+DECLARE_SOA_COLUMN(MDCand1, mDCand1, float);       //! Invariant mass of first candidate as D
+DECLARE_SOA_COLUMN(MDbarCand1, mDbarCand1, float); //! Invariant mass of first candidate as Dbar
+DECLARE_SOA_COLUMN(MDCand2, mDCand2, float);       //! Invariant mass of second candidate as D
+DECLARE_SOA_COLUMN(MDbarCand2, mDbarCand2, float); //! Invariant mass of second candidate as Dbar
+DECLARE_SOA_COLUMN(PairType, pairType, uint8_t);   //! Bitmap with all pair types (DD, DDbar, etc.) a pair of candidates has passed
+DECLARE_SOA_COLUMN(CandidateType1, candidateType1, uint8_t); //! Bitmap with Selected and True info of candidate 1
+DECLARE_SOA_COLUMN(CandidateType2, candidateType2, uint8_t); //! Bitmap with Selected and True info of candidate 2
 // MC info
 DECLARE_SOA_COLUMN(Origin1, origin1, uint8_t);       //! candidate 1 origin
 DECLARE_SOA_COLUMN(Origin2, origin2, uint8_t);       //! candidate 2 origin
@@ -47,31 +47,31 @@ DECLARE_SOA_COLUMN(MatchedMc2, matchedMc2, uint8_t); //! MC matching of candidat
 // Definition of the D meson pair table. Contains the info needed at Data level.
 #define DECLARE_DMESON_PAIR_TABLE(_pair_type_, _marker_value_, _description_)           \
   DECLARE_SOA_TABLE(_pair_type_, "AOD", _description_, o2::soa::Marker<_marker_value_>, \
-                    hf_correlation_d_meson_pair::DeltaPhi,                              \
-                    hf_correlation_d_meson_pair::DeltaEta,                              \
                     hf_correlation_d_meson_pair::PtCand1,                               \
                     hf_correlation_d_meson_pair::PtCand2,                               \
                     hf_correlation_d_meson_pair::YCand1,                                \
                     hf_correlation_d_meson_pair::YCand2,                                \
-                    hf_correlation_d_meson_pair::MCand1,                                \
-                    hf_correlation_d_meson_pair::MCand2,                                \
+                    hf_correlation_d_meson_pair::MDCand1,                               \
+                    hf_correlation_d_meson_pair::MDbarCand1,                            \
+                    hf_correlation_d_meson_pair::MDCand2,                               \
+                    hf_correlation_d_meson_pair::MDbarCand2,                            \
+                    hf_correlation_d_meson_pair::PairType,                              \
                     hf_correlation_d_meson_pair::CandidateType1,                        \
-                    hf_correlation_d_meson_pair::CandidateType2,                        \
-                    hf_correlation_d_meson_pair::DataType);
-
-// definition of the D meson pair table with info at MC level.
-#define DECLARE_DMESON_PAIR_RECOINFO_TABLE(_pair_type_, _marker_value_, _description_)             \
-  DECLARE_SOA_TABLE(_pair_type_, "AOD", _description_ "RECOINFO", o2::soa::Marker<_marker_value_>, \
-                    hf_correlation_d_meson_pair::Origin1,                                          \
-                    hf_correlation_d_meson_pair::Origin2,                                          \
-                    hf_correlation_d_meson_pair::MatchedMc1,                                       \
+                    hf_correlation_d_meson_pair::CandidateType2);
+// Definition of the D meson pair table with info at MC level.
+#define DECLARE_DMESON_PAIR_MCINFO_TABLE(_pair_type_, _marker_value_, _description_)             \
+  DECLARE_SOA_TABLE(_pair_type_, "AOD", _description_ "MCINFO", o2::soa::Marker<_marker_value_>, \
+                    hf_correlation_d_meson_pair::Origin1,                                        \
+                    hf_correlation_d_meson_pair::Origin2,                                        \
+                    hf_correlation_d_meson_pair::MatchedMc1,                                     \
                     hf_correlation_d_meson_pair::MatchedMc2);
 
 // Creation of tables with D Meson Pairs info
-DECLARE_DMESON_PAIR_TABLE(D0Pair, 1, "D0PAIR");                    //! D0 pairs Info
-DECLARE_DMESON_PAIR_RECOINFO_TABLE(D0PairRecoInfo, 1, "D0PAIR");   //! D0 pairs Reconstructed Info
-DECLARE_DMESON_PAIR_TABLE(DplusPair, 2, "DPLUSPAIR");              //! D+ pairs Info
-DECLARE_DMESON_PAIR_RECOINFO_TABLE(DplusPairRecoInfo, 2, "DPLUS"); //! D+ pairs Reconstructed Info
+DECLARE_DMESON_PAIR_TABLE(D0Pair, 1, "D0PAIR");              //! D0 pairs Info
+DECLARE_DMESON_PAIR_MCINFO_TABLE(D0PairMcInfo, 1, "D0PAIR"); //! D0 pairs MC Rec Info
+
+DECLARE_DMESON_PAIR_TABLE(D0PairMcGen, 2, "D0PAIRGEN");            //! D0 pairs MC Gen Kinematic Info
+DECLARE_DMESON_PAIR_MCINFO_TABLE(D0PairMcGenInfo, 2, "D0PAIRGEN"); //! D0 pairs MC Gen Info
 
 } // namespace o2::aod
