+++ PWGHF/HFC/Tasks/CMakeLists.txt
@@ -24,11 +24,6 @@ o2physics_add_dpl_workflow(task-correlation-d-meson-pairs
                     PUBLIC_LINK_LIBRARIES O2Physics::AnalysisCore
                     COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(task-correlation-d-meson-pairs-testing
-                    SOURCES taskCorrelationDMesonPairsTesting.cxx
-                    PUBLIC_LINK_LIBRARIES O2Physics::AnalysisCore
-                    COMPONENT_NAME Analysis)
-
 o2physics_add_dpl_workflow(task-correlation-dplus-hadrons
                     SOURCES taskCorrelationDplusHadrons.cxx
                     PUBLIC_LINK_LIBRARIES O2Physics::AnalysisCore
